a5labs/requirements.txt
ruff
pyyaml
torch == 2.9.1